Move audio top_k tests to the right file and add slow decorator because we load a real model
